requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.1.0
boto3==1.40.19
botocore==1.40.19
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-default-group==1.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-appconf==1.1.0
django-apscheduler==0.7.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-countries-plus==2.2.0
django-filter==25.1
django-jazzmin==3.0.1
django-money==3.5.4
django-phonenumber-field==8.1.0
django-pwa==2.0.1
django-redis==5.4.0
django-timezone-field==7.1
django-tinymce==4.1.0
django-tinymce4-lite==1.8.0
djangorestframework==3.16.0
djangorestframework-csv==3.0.2
dnspython==2.7.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
faker==37.1.0
eventlet==0.39.1
executing==2.2.0
Faker==37.1.0
firebase-admin==6.8.0
flower==2.0.1
frozenlist==1.6.0
future==1.0.0
geek==0.1.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipython==8.32.0
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsmin==3.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
llm==0.21
loguru==0.7.3
lxml==5.4.0
mark==0.10.1
Markdown==3.8
markdownify==0.14.1
marshmallow==3.26.1
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openai==1.77.0
openpyxl==3.1.5
orjson==3.10.18
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
phonenumbers==9.0.4
pillow==11.2.1
pluggy==1.5.0
polars==1.29.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
puremagic==1.29
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pyee==11.1.1
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pymongo==4.12.1
pyotp==2.9.0
pyparsing==3.2.3
pypdf==5.5.0
pyppeteer==2.0.0
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==4.0.1
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
rcssmin==1.1.2
redis==6.0.0
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rjsmin==1.2.2
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.13.1
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.5.0
svglib==1.5.1
tabulate==0.9.0
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.6.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
uritools==5.0.0
urllib3==1.26.20
user-agents==2.2.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
whitenoise==6.9.0
win32_setctime==1.2.0
xhtml2pdf==0.2.17
yarl==1.20.0
zipp==3.21.0
